Renamed files - moreinfo.js is now the correct version for 2.6
